Fix crash when a page action cannot load its icon.
This can happen legitmately if the icon has been moved or
corrupted on disk.

I think we are hitting it too frequently for that to be the
whole explanation, but since this code is moving around a
lot shortly, it seems prudent to put in the quick fix and
worry about what's really going on post-refactor.

BUG=25601
TEST=Remove an extension's page action icon from its folder after installation. Should not crash when attempting to show page action.

Review URL: http://codereview.chromium.org/339005

git-svn-id: http://src.chromium.org/svn/trunk/src@30069 4ff67af0-8c30-449e-8e8b-ad334ec8d88c