Rebaseline LayoutTests/editing/execCommand/indent-paragraphs.html, which was added by the merge.  There's no pixel checked in upstream, but we match except for not showing image placeholders.  The test has the following images:
<img src="http://">

With GURL, the URL is rejected and it doesn't try to load the image.
In KURL, the URL is "valid", but it fails to load an image so the place
holder image is shown instead.  Since that's not what the test is trying
to test, just rebaseline.

Review URL: http://codereview.chromium.org/342006

git-svn-id: http://src.chromium.org/svn/trunk/src@30283 4ff67af0-8c30-449e-8e8b-ad334ec8d88c