Clobber layout test results at the beginning of each run. This should fix issues with the layout_test_results.zip files on the bots have results from other runs, which makes debugging flakiness and other failures harder.
Add a flag, clobber-old-results that defaults to false.
We'll need to add the flag to the bots, but this will
maintain the old behavior for local runs of run_webkit_tests.

I'd also be fine not having a flag and always clobbering,
but I don't know if people rely on the current behavior
for their workflows.
Review URL: http://codereview.chromium.org/340064

git-svn-id: http://src.chromium.org/svn/trunk/src@30736 4ff67af0-8c30-449e-8e8b-ad334ec8d88c