Rewrite of chrome.exe startup code
A lot of cruft and repeated code has deposited over the years on chrome's initialization code.
This CL makes it all much more clear and straightforward. There is no fundamental change of
behavior except the order of certain things is different but it should not alter the observed
operation.

- chrome's and chromium load is fundamentally the same but most of the code was repeated
- chrome's way to load the dll was incorrect: using a relative path with LOAD_WITH_ALTERED_SEARCH_PATH
- Use of SearchPath() was dangerous and not needed
- removed google_update_client.cc and .h
- removed bunch of #ifdefs

TEST=all convered by UI tests already except [1]
BUG=none

[1] The only thing I don't see convered by test is the restart dialog ('woa! chrome crashed').


Review URL: http://codereview.chromium.org/345036

git-svn-id: http://src.chromium.org/svn/trunk/src@30934 4ff67af0-8c30-449e-8e8b-ad334ec8d88c