GTK: fix certain corner cases where overflow chevron wouldn't show.
When a tool item is added but there is no room for it, it doesn't get an allocation right away, so our check to see if it fit was failing. This would be fixed the next time there was an allocate event (e.g. due to a resize).

I noticed while debugging this that SetChevronState is getting called like 10 times in a row when you type in a url and press enter. This is in repsonse to allocate signals. I don't know why there are so many allocate signals, but it warrants further investigation.

BUG=26736

Review URL: http://codereview.chromium.org/372002

git-svn-id: http://src.chromium.org/svn/trunk/src@31147 4ff67af0-8c30-449e-8e8b-ad334ec8d88c