Fixes navigations chrome-internal: to actually show the NTP.
We now rewrite chrome-internal: (the old NTP URL) to chrome://newtab.  Before, we were not properly showing the NTP for this URL, which caused some problems with further navigations (particularly form submissions) from that tab.

Also, the default kHomePage string is now chrome://newtab instead of the outdated chrome-internal:.  The logic in GeneralPageView is already designed to not show this URL in the preferences dialog.

BUG=6564
TEST=NewTabUITest.ChromeInternalLoadsNTP

Review URL: http://codereview.chromium.org/363019

git-svn-id: http://src.chromium.org/svn/trunk/src@31182 4ff67af0-8c30-449e-8e8b-ad334ec8d88c