GTK: Change find bar to use a floating container rather than some hacked up gtkfixed business.
The straw that broke this camel's back was when I realized that showing the findbar in fullscreen mode would inescapably mean an extra 1 pixel border at the top.

Also fix z-ordering problems for findbar that have been around since I introduced the floating bookmark bar (the fact that no one filed a bug on this suggests that it's not that noticeable).

BUG=none
TEST=find bar still works. find bar works in fullscreen. the toolbar bottom border still exists at all times. there are no z-ordering problems. the find bar still gets out of way of find results.

Review URL: http://codereview.chromium.org/373007

git-svn-id: http://src.chromium.org/svn/trunk/src@31296 4ff67af0-8c30-449e-8e8b-ad334ec8d88c