Add tests that pass on all platforms to the updates JSON for autoupdating test_expectations. For those tests, we just say that the expectations they're currently listed with are "extra".
Also reduce the MIN_RUNS_FOR_FLAKE to 2. If a test has the same results
twice in a row, assume it's not flakiness. In the cases where a test
is so flaky that sometimes it has the same results twice in a row, it'll 
show up as flakiness due to other runs.
Review URL: http://codereview.chromium.org/393018

git-svn-id: http://src.chromium.org/svn/trunk/src@31972 4ff67af0-8c30-449e-8e8b-ad334ec8d88c