This CL fixes the document-open.html and window-open.html plugin layout tests. These were landed in webkit as fixes for bug https://bugs.webkit.org/show_bug.cgi?id=31067
The layout test plugin changes are on similar lines as the version upstream.

We need to rebaseline the expectations as the line numbers differ in test shell and dumprendertree.

The document-open.html crashes in test_shell as this test causes the existing document to be taken down
which takes down the plugin instance and the associated plugin stream. This works in the webkit plugin
implementation as they grab a reference on the plugin stream before calling into the plugin to protect
from the stream being destroyed.

Our fix is on similar lines.

Bug=27164

Review URL: http://codereview.chromium.org/393017

git-svn-id: http://src.chromium.org/svn/trunk/src@31991 4ff67af0-8c30-449e-8e8b-ad334ec8d88c