Fix find bar scrolling with the page after switching back to tab.
We need to do a BrowserView::Layout() (which updates things like the bookmarks bar and the find bar) both in between the TabContents detach/attach AND after the final attach. This way, we toggle the bookmarks bar at the right time, but we still end up setting the find bar properly.

R=beng
BUG=27960
TEST=Navigate to a page where you can see the vertical scroll bar. Press Ctrl+F to open the find bar. Switch to another tab and then back to the original tab. Scroll down/up the page. The find bar should stay in place, not scroll with the page or disappear.

Review URL: http://codereview.chromium.org/397036

git-svn-id: http://src.chromium.org/svn/trunk/src@32284 4ff67af0-8c30-449e-8e8b-ad334ec8d88c