gtk: Get rid of status bubble flicker.
The bubble was appearing in its default location
when getting shown while the mouse was nearby; it
would move to the correct location on the next
motion event.

I think that there's still a flicker in the wrong
location when closing the download bar and then
triggering the bubble, but the fix is non-obvious
to me.  It's possible that it's not noticeable on
anything faster than the slow DSL line that I'm
using to run Chrome remotely to test this.

BUG=28494
TEST=repeated steps from bug and checked that download bar behavior doesn't appear to have regressed either

Review URL: http://codereview.chromium.org/431007

git-svn-id: http://src.chromium.org/svn/trunk/src@32789 4ff67af0-8c30-449e-8e8b-ad334ec8d88c