Take 2 at moving NTP CSS resources into the NTP resource cache.
Moving the HTML to the cache saved 8ms on New Tab Warm on Windows,
maybe we can shave a few more ms off by moving the CSS.

The last change had a regression on the mac new tab perf, so I suspect
it's the extra round trip to the UI thread when trying to load the CSS
(the UI thread is probably just busy at that time).  Instead, I'm moving
the generation of the CSS to the time the object is constructed, rather
than when the resource is requested.  This is like the code before,
except sometimes we'll hit the cache.

BUG=26228

Review URL: http://codereview.chromium.org/427003

git-svn-id: http://src.chromium.org/svn/trunk/src@32815 4ff67af0-8c30-449e-8e8b-ad334ec8d88c