Fix a flaky dev tools sanity test.
The test only works if the source of scripts are found in a specific
order.  This cannot be ensured since the source is found by traversing
the heap.  The scripts tab is only visible after processing the
debugger test page, so only check for visibility in that case.

Enable test on Linux again where it passes after this change.

TEST=none
BUG=26540
Review URL: http://codereview.chromium.org/434044

git-svn-id: http://src.chromium.org/svn/trunk/src@32968 4ff67af0-8c30-449e-8e8b-ad334ec8d88c