Fixes the compiled o3djs libraries.
The issue was that lineprimitves.js creates
an object that inherits from something in
primitives.js but because it appeared first
in the compiled version the inherit command
failed since the object it inherited from
was not yet defined.

It might be possible to make build_docs.py
move all the inherit commands to the end
of the compiled o3djs libs. I'll have to 
look into it.

In the meantime here is a manual fix.

Review URL: http://codereview.chromium.org/464030

git-svn-id: http://src.chromium.org/svn/trunk/src@33773 4ff67af0-8c30-449e-8e8b-ad334ec8d88c