Make BrowserProcess::GetApplicationLocale thread safe and migrate callers of l10n_util::GetApplicationLocale to use this instead.
In the browser process, it's wrong to call
l10n_util::GetApplicationLocale with an empty string because then
it won't consider the user pref value when resolving the locale.

On Linux, it's also wrong to call l10n_util::GetApplicationLocale
after startup because the call touches disk and on Linux, we assume
that all of the program files can be deleted after startup (so
updates in place can work).

Review URL: http://codereview.chromium.org/476002

git-svn-id: http://src.chromium.org/svn/trunk/src@34206 4ff67af0-8c30-449e-8e8b-ad334ec8d88c