Fix radio items in views::Menu2 for Linux.
Suppose there are three radio items A, B, C, and A is now being
checked. If you click C, commands associated with A and C are both
executed. This is because A also receives the "activate" signal,
before C receives it. We want only the command associated with A
to get executed.

BUG=30310
TEST=try; manually with view_examples

Review URL: http://codereview.chromium.org/491076

git-svn-id: http://src.chromium.org/svn/trunk/src@34644 4ff67af0-8c30-449e-8e8b-ad334ec8d88c