Convert webkit_glue.cc over to using WebKit API.
Removed unnecessary code to reset the JS line number, which was
only added for the Chrome+JSC build.  I don't have a way to test
if it would still be necessary, and I'd rather not add a WebKit
API for it unless I can prove that I need it.

R=dglazkov
BUG=28065
TEST=none

Review URL: http://codereview.chromium.org/517038

git-svn-id: http://src.chromium.org/svn/trunk/src@35590 4ff67af0-8c30-449e-8e8b-ad334ec8d88c