Use no_strict_aliasing=1 by default.
This will turn on -fno-strict-aliasing. There are regular aliasing problems
in the codebase and at least one of these was not caught by -Wall -Werror
so defaulting to -fno-strict-aliasing seems like a good idea especially
given the increased popularity of gcc 4.4 with it's use in karmic, fedora and in
chromeos chroot builds et. al.

The long term plan is to have some test infrastructure in place that will test
4.4 binaries and to fix the aliasing issues.

BUG=28749

Review URL: http://codereview.chromium.org/519034

git-svn-id: http://src.chromium.org/svn/trunk/src@36151 4ff67af0-8c30-449e-8e8b-ad334ec8d88c