We no longer need to (or should) specify "LayoutTests/" in front of any layout tests files or test names. This brings us into conformance w/ upstream webkit.
Eventually all the directories under
webkit/data/layout_tests/platform/chromium-{mac,win*,linux}/LayoutTests need
to move up a directory to also be compatible with WebKit. That will be a
separate changelist since it is a rote rename of thousands of files. In the
meantime we modify the baseline search path to look in both chromium-X and
then chromium-X/LayoutTests .

  BUG=31567
  R=ojan@chromium.org
  TEST=none

Review URL: http://codereview.chromium.org/545038

git-svn-id: http://src.chromium.org/svn/trunk/src@36214 4ff67af0-8c30-449e-8e8b-ad334ec8d88c