Remove the fallback Mozilla code for parsing FTP LIST response.
The new parser seems to be compatible enough to do that. The Mozilla code was very
helpful in the process of developing the new parser.

Also add UI encouraging users to submit bug reports when we can't parse the listings,
and an option to see the raw data sent by the server. This should allow us to fix
remaining compatibility problems with very rare listing types or variations.

When ?raw is found at the end of an FTP url and it is a directory listing, the parsing logic is bypassed and the data is displayed as-is with text/plain MIME type.

TEST=none
BUG=25520

Review URL: http://codereview.chromium.org/549053

git-svn-id: http://src.chromium.org/svn/trunk/src@36632 4ff67af0-8c30-449e-8e8b-ad334ec8d88c