GTK: Use color data from the MetaFrames class in gtk+ theme mode.
Most themes have specific data for a "MetaFrames" class that we weren't using. This fixes the titlebar color
for the xfce themes. I suspect that this also fixes 32173, but I don't have
a Karmic machine yet so I can't test.

BUG=32173
TEST=Themes with different titlebar and selection colors (Xfce-b5 was what I
tested with) shouldn't use the selection color for the titlebar.

Review URL: http://codereview.chromium.org/552092

git-svn-id: http://src.chromium.org/svn/trunk/src@36867 4ff67af0-8c30-449e-8e8b-ad334ec8d88c