remove ICU includes from l10n_util.h
95% of users of l10n_util use it for some functions; the other 5%
want some complicated templates that pull in a ton of ICU headers
as well.

Before this change, the average includer of l10n_util.h pulled in
an additional 80 subheaders because of it.

Additionally, #including ICU headers from a header makes the includee
depend on having the ICU include path in the -I header.

Review URL: http://codereview.chromium.org/515059

git-svn-id: http://src.chromium.org/svn/trunk/src@37032 4ff67af0-8c30-449e-8e8b-ad334ec8d88c