Clean up bookmark manager extension api code a bit.
The macros used to verify the extension adds a return false but our function used to return a BookmarkNode*. I now changed this to a helper funciton that returns a pointer and NULL in the case of failure.

I also added a GetBookmarkNodesFromArguments which returns a vector.

BUG=None
TEST=Bookmark manager copy/paste should still work.

Review URL: http://codereview.chromium.org/561024

git-svn-id: http://src.chromium.org/svn/trunk/src@37911 4ff67af0-8c30-449e-8e8b-ad334ec8d88c