Implement launch disposition for extension-apps.
This change adds an --app-id command switch that signifies that the extension-app with the given id should be launched according to its configuration.

It also adds parsing for app.window_type in the manifest and the behavior that when installed and a desktop shortcut is created, the --app-id switch is used rather than the --app switch.

BUG=32361

Review URL: http://codereview.chromium.org/573016

git-svn-id: http://src.chromium.org/svn/trunk/src@38184 4ff67af0-8c30-449e-8e8b-ad334ec8d88c