Fix for bug 26397: Bookmark manager in Incognito mode doesn't show Sync Bookmarks button
The bug exists because the BookmarkManagerView checks the profile to see if it has a ProfileSyncService set up to see whether it should show the "Sync Bookmarks" button.  Incognito profiles do not have a ProfileSyncService, even if the corresponding "original profile" has a PSS.  So the fix is to check the original profile (here called "profile_"), not the profile.

This is already fixed in Linux (bookmark_manager_gtk.cc).  Also, all native versions are scheduled to be replaced by a common bookmark manager soon.

BUG=26397
TEST=manual

Review URL: http://codereview.chromium.org/574022

git-svn-id: http://src.chromium.org/svn/trunk/src@38625 4ff67af0-8c30-449e-8e8b-ad334ec8d88c