Attempt to fix the Linux Views and ChromiumOS builds following r38658.
chrome/browser/views/tab_contents/render_view_context_menu_win.cc appears to
be a views file, not a Windows file, so why is it named with _win.cc?

This is a speculative change.  It may not work, in which case the file will
need to be renamed.

BUG=none
TEST=Linux Builder (Views dbg) and Linux Builder (ChromiumOS) compile steps
     turn green
Review URL: http://codereview.chromium.org/598042

git-svn-id: http://src.chromium.org/svn/trunk/src@38667 4ff67af0-8c30-449e-8e8b-ad334ec8d88c