Added factories for HttpAuthHandler.
The driving rationale for this change was to prevent choosing an AuthHandler when it
is not supported on the system due to a missing runtime component (such as not being
able to locate a gssapi shared library when seeing a Negotiate scheme).

It also has the advantage (currently unused) of determining some per-auth-scheme properties
only the first time that a challenge for that scheme is seen (such as maximum token length for
the SSPI implementation of NTLM).

Finally, it may make unit tests easier to generate since the factory can be easily mocked.

BUG=34795
TEST=New unit test for HttpAuthHandlerDispatchFactory.

Review URL: http://codereview.chromium.org/582007

git-svn-id: http://src.chromium.org/svn/trunk/src@39065 4ff67af0-8c30-449e-8e8b-ad334ec8d88c