Circular relationships between .gyp files should be errors.  Make them errors, but currently only on the Mac.
These relationships should be errors on all platforms, but some currently
exist on non-Mac platforms.  See http://crbug.com/35878.  Because the Mac is
the only platform where a circular dependency between .gyp files is known to
cause tangible problems, the portions of Chromium's .gyp files that are used
by Macs have been fixed to remove these relationships, and the check is left
enabled on the Mac to ensure that no new ones are created.

BUG=35308
TEST=none
Review URL: http://codereview.chromium.org/600151

git-svn-id: http://src.chromium.org/svn/trunk/src@39128 4ff67af0-8c30-449e-8e8b-ad334ec8d88c