Temporary fix for issue 34061.
Change that signals that drag-drop operations are immediately terminated in ExtensionHost views.  Previously, the view would become unresponsive, because the RenderViewHost would remain 'stuck' in the dragging logic, and all input events would be disregarded.

This fixes the issue described in the bug, but the long-term solution is to properly implement drag-drop operations for extension-views.

BUG=34061.
TEST=None.

Review URL: http://codereview.chromium.org/650059

git-svn-id: http://src.chromium.org/svn/trunk/src@39627 4ff67af0-8c30-449e-8e8b-ad334ec8d88c