Make the translation bar not show up again when you closed it in a page and navigate in page. This is the case with web apps like GMail where many actions result in an in-page navigation. It would originally bring back the translation infobar, which is very anoying.
This CL fixes the is_in_page state of LoadCommittedDetails, it was always false. 
This triggered a bug in the SSLManager that was wrongly not setting the SSL states when an in-page navigation is performed. 
Since we create a new navigation entry for in-page navigations, not setting these states would make a bad page being reported as OK after an in-page navigation. 

See original review:
http://codereview.chromium.org/650146/show

BUG=36304
TEST=Login to gmail, change the GMail language to a language different than 
the Chrome language. When a translate infobar shows up, close it. 
Click on the different links in GMail (inbox, drafts, select email...) 
the translate infobar should not show again. 

Review URL: http://codereview.chromium.org/650207

git-svn-id: http://src.chromium.org/svn/trunk/src@39700 4ff67af0-8c30-449e-8e8b-ad334ec8d88c