Make the pepper 2D flush callback actually function as advertised. It will now get called asynchronously when the bits are actually copied to the screen, rather than synchronously from inside the paint function. This makes it useful for plugins to use the callback for rate limiting.
This also adds a lot of infrastructure for running tests on pepper devices, and
includes a unit test for the new flush behavior.

I made the existing RenderProcess object an abstract interface and made the
existing MockProcess (renamed to be more clear) implement that. This avoids
a static cast that would actually crash during a unit test because some code
was hardcoded to expect a RenderProcess object.

This fixes base's IDMap iterator which has apparently never been used for an
IDMap with ownership semantics.

TEST=Unit test included
BUG=none
Review URL: http://codereview.chromium.org/661124

git-svn-id: http://src.chromium.org/svn/trunk/src@40490 4ff67af0-8c30-449e-8e8b-ad334ec8d88c