linux (sysroot): force prefix=/usr in pkg-config
This lets us use portage-based cross-compiled sysroots, that force the prefix to be an absolute path within the portage chroot.
In particular that lets us build the ARM sysroot using the Chrome OS build system.
Ubuntu-based sysroots already have prefix=/usr, so this is a no-op in that case.

Review URL: http://codereview.chromium.org/669139

git-svn-id: http://src.chromium.org/svn/trunk/src@40712 4ff67af0-8c30-449e-8e8b-ad334ec8d88c