Fix crashing in FFmpegGlueTest.Write and friends
BUG=36037
TEST=FFmpegGlueTest.* runs on build bot

FFmpegGlueTest.Write was crashing on build bots due to that it
has dependencies on previous test to run. And strangely the build bot
didn't run all the test cases, resulting the crash.

This is fixed by extracting the initialization step and putting it
into a common test setup task.

Review URL: http://codereview.chromium.org/668034

git-svn-id: http://src.chromium.org/svn/trunk/src@40800 4ff67af0-8c30-449e-8e8b-ad334ec8d88c