First cut at custom user style sheets.
Enabled with the --enable-user-stylesheet flag which
causes chrome to read
<user-data-dir>/<profile>/User StyleSheet/Custom.css
at startup and set it as the user style sheet.

This version never reloads the user style sheet, I'll
have to bring back FileWatcher for that.  I also put the user
stylesheet in a subdir because the implementation of
FileWatcher will watch the parent dir (this is what the OS
apis give me) and watching the profile dir will cause
lots of activity.

BUG=2393

Review URL: http://codereview.chromium.org/660349

git-svn-id: http://src.chromium.org/svn/trunk/src@40882 4ff67af0-8c30-449e-8e8b-ad334ec8d88c