Fix some issues with text size calculation and message boxes.
First, size calculations for multi-line, character-breaking strings with no prespecified width were returning useless values, and when given pathological input, could take extremely long time.

Second, the flag calculations for text layout were a bit odd.  For example, we wouldn't add an ending ellipsis on elided text if it was multi-line, and we wouldn't add an ellipsis on a clipped word if character breaking was off.  Fixed these and reordered the calculations to match the order in which the header declares the flags.

Third, message boxes shouldn't need to trim their inputs, or add an ellipsis character, as both of these should now be handled safely by the underlying code.

BUG=34721
TEST=alert("AAAAA") for strings of As that are, say, 4000 characters long should not cause the UI to hang up for noticeable lengths of time
Review URL: http://codereview.chromium.org/810003

git-svn-id: http://src.chromium.org/svn/trunk/src@41343 4ff67af0-8c30-449e-8e8b-ad334ec8d88c