When ChromeFrame switches to IE on receiving the OnHttpEquiv notification indicating the presence of a meta tag indicating that the page is to be rendered in Chrome, we check if the notification is received for a site rendered in an IFrame to ensure that we don't switch in this case. This code is not reliable and we end up assuming that a top level navigation is actually an embedded frame.
To work around this we now maintain a pending navigation count in BeforeNavigate2, which is decremented in
NavigateComplete2 and NavigateError. We perform the check for an embedded document only if the pending navigation
count is greater than 1.

This fixes http://code.google.com/p/chromium/issues/detail?id=36825

The other change made is to add a delay of 100ms in SendString between each character to better reflect actual
timing for real user input.

Bug=36825

Review URL: http://codereview.chromium.org/837008

git-svn-id: http://src.chromium.org/svn/trunk/src@41463 4ff67af0-8c30-449e-8e8b-ad334ec8d88c