Initial support for inspecting extension popups.
The primary change in this CL is a refactor which makes ExtensionPopup a bit more self-contained WRT its clients. It adds the ability to specify an "inspect_with_devtools" flag to its Show() method which will cause the popup to remain open regardless of losing focus and to focus a devtools window on the popup's render view host.

This CL also pulls apart some aspects of the extension_popup_api from ExtensionFunctionDispatcher and ExtensionHost.

Still remaining to be done are:

1) Also the popup to stay open when the host window drags (it current closes)
2) Support for GTK
3) Support for Mac

BUG=24477

Review URL: http://codereview.chromium.org/1001002

git-svn-id: http://src.chromium.org/svn/trunk/src@41854 4ff67af0-8c30-449e-8e8b-ad334ec8d88c